logo change and link adjustment, also moved file placement
