+++ WIT-website/src/components/NavBar.jsx
@@ -1,29 +1,30 @@
 import React from 'react';
-import './Styles/NavBar.css';
+import '../Styles/NavBar.css';
+import logo from '../assets/Logo/Women_In_Tech_Logo_pink.png'; // Use import for consistent path resolution
 
 const Navbar = () => {
     return (
         <nav className="navBar">
             <div className="navbar-leftside">
                 <a href="/" className="logo">
-                <img 
-                    src="src/assets/Logo/Women_In_Tech_Logo.png"
-                    className="wit-logo"
-                />
-                <span className="pink-dot">.</span>
+                    <img
+                        src={logo}
+                        alt="Women in Tech Logo"
+                        className="wit-logo"
+                    />
                 </a>
             </div>
 
             <div className="navbar-right">
                 <ul className="nav-links">
-                    <li><a href="/home" className="nav-link">Home</a></li>
+                    <li><a href="/" className="nav-link">Home</a></li>
                     <li className="nav-item">
                         <div className="dropdown">
-                            <button className='dropbutton'>
+                            <button className="dropbutton">
                                 About
-                                <i className='caret-down'></i>
+                                <i className="caret-down"></i>
                             </button>
-                            <div className='dropdown-content'>
+                            <div className="dropdown-content">
                                 <a href="/about/aboutUs">About Us</a>
                                 <a href="/about/team">Meet Our Team!</a>
                             </div>
@@ -39,3 +40,4 @@ const Navbar = () => {
 };
 
 export default Navbar;
+
